requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
google-generativeai==0.3.2
yookassa==2.3.5
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0 